refactor(StringExpression at el): Refactor parseToMathExpression function to StringExpression object
